The JavaScript for the chat client workings. Still contains debugging code. (#8)
#### How it works
Using the powers of a socket to send and receive messages.

#### Testing
There's lots of debugging code still included. Pressing `w` allows for receiving a message, and scrolling up `pastes` ten example messages to the top of the messages list.

NB!! There is also template code in the server-side, without which the client-side code wouldn't work, as it would have no socket to connect to. That code shall be removed and replaced later on.

Co-authored-by: Reimann Robert <reimannrobert707@gmail.com>
Co-authored-by: Kanguste <reimannrobert707@gmail.com>
Reviewed-on: https://git.01.kood.tech/Kanguste/real-time-forum/pulls/8
Co-authored-by: Robert Reimann <kanguste@noreply.git.01.kood.tech>
Co-committed-by: Robert Reimann <kanguste@noreply.git.01.kood.tech>